pyproject.toml
[project]
name = "ffmpeg-ws"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "beautifulsoup4>=4.12",
    "fake-useragent==1.4.0",
    "faster-whisper>=1.0.3",
    "matplotlib==3.7.2",
    "moviepy>=1.0.3",
    "numpy==1.24.3",
    "onnxruntime>=1.15.0,<1.20.0",
    "opencc-python-reimplemented>=0.1.7",
    "pillow==10.0.0",
    "pyqt5>=5.15",
    "requests>=2.32",
    "selenium>=4.22",
    "speechrecognition>=3.10.4",
    "tokenizers==0.13.3",
    "tqdm==4.66.1",
    "vosk>=0.3.45",
    "webdriver-manager>=4.0",
]
